refactor:Sheriff remove option int cast

+++ ExtremeRoles/Roles/Solo/Crewmate/Sheriff.cs
@@ -125,21 +125,21 @@ namespace ExtremeRoles.Roles.Solo.Crewmate
             CustomOptionBase parentOps)
         {
             CustomOption.Create(
-                GetRoleOptionId((int)SheriffOption.CanShootAssassin),
+                GetRoleOptionId(SheriffOption.CanShootAssassin),
                 string.Concat(
                     this.RoleName,
                     SheriffOption.CanShootAssassin.ToString()),
                 false, parentOps);
 
             CustomOption.Create(
-                GetRoleOptionId((int)SheriffOption.CanShootNeutral),
+                GetRoleOptionId(SheriffOption.CanShootNeutral),
                 string.Concat(
                     this.RoleName,
                     SheriffOption.CanShootNeutral.ToString()),
                 true, parentOps);
 
             CustomOption.Create(
-                GetRoleOptionId((int)SheriffOption.ShootNum),
+                GetRoleOptionId(SheriffOption.ShootNum),
                 string.Concat(
                     this.RoleName,
                     SheriffOption.ShootNum.ToString()),
@@ -150,11 +150,11 @@ namespace ExtremeRoles.Roles.Solo.Crewmate
         protected override void RoleSpecificInit()
         {
             this.shootNum = OptionHolder.AllOption[
-                GetRoleOptionId((int)SheriffOption.ShootNum)].GetValue();
+                GetRoleOptionId(SheriffOption.ShootNum)].GetValue();
             this.canShootNeutral = OptionHolder.AllOption[
-                GetRoleOptionId((int)SheriffOption.CanShootNeutral)].GetValue();
+                GetRoleOptionId(SheriffOption.CanShootNeutral)].GetValue();
             this.canShootAssassin = OptionHolder.AllOption[
-                GetRoleOptionId((int)SheriffOption.CanShootAssassin)].GetValue();
+                GetRoleOptionId(SheriffOption.CanShootAssassin)].GetValue();
             this.killCountText = null;
         }
